:hammer: refactor: tournament 폴더 관련 import 수정
