+++ src/main/java/com/gg/server/domain/tournament/controller/TournamentController.java
@@ -1,2 +1,11 @@
-package com.gg.server.domain.tournament.controller;public class TournamentController {
+package com.gg.server.domain.tournament.controller;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/pingpong/tournament")
+public class TournamentController {
 }
