+++ src/main/java/com/gg/server/domain/tournament/data/TournamentGame.java
@@ -1,7 +1,7 @@
-package com.gg.server.domain.touranment.data;
+package com.gg.server.domain.tournament.data;
 
 import com.gg.server.domain.game.data.Game;
-import com.gg.server.domain.touranment.type.TournamentRound;
+import com.gg.server.domain.tournament.type.TournamentRound;
 import com.gg.server.global.utils.BaseTimeEntity;
 import javax.persistence.Column;
 import javax.persistence.Entity;
