+++ src/main/java/com/gg/server/domain/tournament/data/TournamentRepository.java
@@ -1,6 +1,5 @@
-package com.gg.server.domain.touranment.data;
+package com.gg.server.domain.tournament.data;
 
-import com.gg.server.domain.game.data.Game;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TournamentRepository extends JpaRepository<Tournament, Long> {
