+++ src/main/java/com/gg/server/domain/tournament/data/TournamentUserRepository.java
@@ -1,4 +1,4 @@
-package com.gg.server.domain.touranment.data;
+package com.gg.server.domain.tournament.data;
 
 import org.springframework.data.jpa.repository.JpaRepository;
