+++ src/main/java/com/gg/server/domain/tournament/service/TournamentService.java
@@ -1,2 +1,9 @@
-package com.gg.server.domain.tournament.service;public class TournamentService {
+package com.gg.server.domain.tournament.service;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class TournamentService {
 }
